refactor: move legacy descriptor prop rendering into its own file
no semantic changes
