+++ .changeset/calm-bobcats-grow.md
@@ -1,0 +1,5 @@
+---
+"@makeswift/runtime": patch
+---
+
+refactor: move legacy descriptor prop rendering into its own file
